backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
supabase==2.0.0
redis==5.0.1
openai==1.3.0
kerykeion==4.3.0
pyswisseph==2.10.3.2
pytz==2022.7.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
greenlet==3.0.3
# VedAstro integration for comprehensive Vedic astrology calculations
vedastro>=1.0.0
